+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [HexaPDF::Document::Annotations], accessible via
   [HexaPDF::Document#annotations], as convenience interface for working with
   annotations
+* [HexaPDF::Type::Annotations::AppearanceGenerator] as central class for
+  generating appearance streams
 * [HexaPDF::Type::Annotations::Line] for line annotations
 * [HexaPDF::Type::Annotation#opacity] for setting the opacity values when
   regenerating the appearance stream